requirements.txt
# ASGI support for Django, provides asynchronous server gateway interface utilities
asgiref==3.8.1

# Django web framework, core of the project
Django==5.1.7

# Enables automatic browser reload during development for faster feedback
django-browser-reload==1.18.0

# Integrates Tailwind CSS with Django for modern, utility-first styling
django-tailwind==3.8.0

# Library for processing XML and HTML documents efficiently
lxml==6.0.1

# Fundamental package for numerical computing in Python
numpy==2.3.2

# Powerful data analysis and manipulation library
pandas==2.3.1

# Library for reading, writing, and manipulating PDF files
PyPDF2==3.0.1

# Provides robust date parsing and extensions to the standard datetime module
python-dateutil==2.9.0.post0

# Library for creating and updating Microsoft Word (.docx) files
python-docx==1.2.0

# Timezone definitions for accurate date/time calculations
pytz==2025.2

# Python 2 and 3 compatibility utilities
six==1.17.0

# SQL parsing library used internally by Django for query management
sqlparse==0.5.3

# Adds extra type hints and extensions to Python typing system
typing_extensions==4.14.1

# Timezone data required for proper timezone handling
tzdata==2025.2

# MySQL client library for Python, enables database connections using MySQLdb
mysqlclient==2.2.0

# Pure Python MySQL client, alternative to mysqlclient
PyMySQL>=1.0.2

# HTTP library for making requests to external APIs or websites

requests 

# Library for interacting with the load testing tool
psutil
beautifulsoup4

